backends/drm: support CTM for simple color transformations
In order to work around hardware and drivers that aren't capable of applying
a LUT, calculate a factor for each color channel and apply it with the CTM if
available. This doesn't result in usable results for display calibration but it
works well enough for night color and software brightness.

BUG: 455720